services/rule_engine/requirements.txt
kafka-python==2.0.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
